requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
pytelegrambotapi==4.3.1
python-slugify==5.0.2
pendulum==2.1.2
emoji==1.6.3